add -h and optional list of buckets to size
By default, it gets a list of all buckets in your partition  (rclone lds lts:/) and then gets size on each of them.

Reminder: you must set up the rclone "lts" endpoint before running this!

Since you can have access to buckets on that endpoint that are not in your partition, and there is no way to query that list, added support for sizing arbitrary buckets so you can provide the name(s) of one you have access to, and it will size thoses.